separated mmvtkm
