+++ plugins/mmvtkm/src/mmvtkm.cpp
@@ -9,8 +9,6 @@
 
 #include "mmvtkm/mmvtkmDataSource.h"
 #include "mmvtkm/mmvtkmFileLoader.h"
-#include "mmvtkm/mmvtkmMeshRenderTasks.h"
-//#include "mmvtkm/mmvtkmRenderer.h"
 #include "mmvtkm/mmvtkmDataCall.h"
 #include "mmvtkm/mmvtkmStreamLines.h"
 
@@ -31,8 +29,6 @@ public:
         // register modules
         this->module_descriptions.RegisterAutoDescription<megamol::mmvtkm::mmvtkmDataSource>();
         this->module_descriptions.RegisterAutoDescription<megamol::mmvtkm::mmvtkmFileLoader>();
-        this->module_descriptions.RegisterAutoDescription<megamol::mmvtkm::mmvtkmMeshRenderTasks>();
-        // this->module_descriptions.RegisterAutoDescription<megamol::mmvtkm::mmvtkmDataRenderer>();
         this->module_descriptions.RegisterAutoDescription<megamol::mmvtkm::mmvtkmStreamLines>();
 
         // register calls
