+++ plugins/mmvtkm_gl/shaders/vtkm_gl/streamline_fragment.glsl
