+++ plugins/mmvtkm_gl/src/stdafx.cpp
@@ -1,0 +1,15 @@
+/*
+ * stdafx.cpp
+ * Copyright (C) 2006-2015 by MegaMol Team
+ * Alle Rechte vorbehalten.
+ */
+
+#include "stdafx.h"
+
+/*
+ * Intentionally empty
+ *
+ * reference any additional headers you need in STDAFX.H
+ * and not in this file
+ */
+
